+++ cookbooks/jobs/recipes/backup.rb
@@ -20,17 +20,17 @@
 
 
 box  = node[:box]
-jobs = box['jobs']
+jobs = box[:jobs]
 
-jobs['users'].each do |username|
-  usr = box['users'][username]
+jobs[:users].each do |username|
+  usr = box[:users][username]
 
-  jobs_dir = "#{usr['home']}/#{jobs['folder']}"
+  jobs_dir = "#{usr[:home]}/#{jobs[:folder]}"
   backup_job_dir = "#{jobs_dir}/backup"
 
   directory backup_job_dir do
     owner username
-    group usr['group']
+    group usr[:group]
     mode 00755
   end
 
@@ -38,7 +38,7 @@ jobs['users'].each do |username|
   template "#{backup_job_dir}/backup.job" do
     source "/jobs/backup/backup.job.erb"
     owner username
-    group usr['group']
+    group usr[:group]
     mode 00755
     backup false
     variables(
@@ -47,18 +47,18 @@ jobs['users'].each do |username|
   end
 
 
-  backup_script = "#{backup_job_dir}/backup-#{box['name']}"
-  backup_tasks  = jobs['profiles'][username]['backup']['tasks']
+  backup_script = "#{backup_job_dir}/backup-#{box[:name]}"
+  backup_tasks  = jobs[:profiles][username][:backup][:tasks]
 
   template backup_script do
     source "/jobs/backup/backup-box_name.erb"
     owner username
-    group usr['group']
+    group usr[:group]
     mode 00755
     backup false
     variables(
       :backup_job_path => backup_job_dir,
-      :box_name        => box['name'],
+      :box_name        => box[:name],
       :jobs_setup_file => "#{jobs_dir}/setup",
       :backup_tasks    => backup_tasks
     )
@@ -74,62 +74,62 @@ jobs['users'].each do |username|
 
 
   backup_tasks.each do |task|
-    task_drive = task['drive'].split(' ').map(&:downcase).join('_')
+    task_drive = task[:drive].split(' ').map(&:downcase).join('_')
     backup_task_dir = "#{backup_job_dir}/#{task_drive}"
 
     directory backup_task_dir do
       owner username
-      group usr['group']
+      group usr[:group]
       mode 00755
     end
 
-    template "#{backup_task_dir}/#{task['name']}" do
+    template "#{backup_task_dir}/#{task[:name]}" do
       source "/jobs/backup/drive/backup-drive.erb"
       owner username
-      group usr['group']
+      group usr[:group]
       mode 00755
       backup false
       variables(
         :backup_task_dir     => backup_task_dir,
-        :backup_task_name    => task['name'],
-        :backup_drive        => task['drive'],
-        :backup_task_folder  => task['folder'],
+        :backup_task_name    => task[:name],
+        :backup_drive        => task[:drive],
+        :backup_task_folder  => task[:folder],
         :backup_task_drive   => task_drive,
-        :backup_task_targets => task['targets']
+        :backup_task_targets => task[:targets]
       )
     end
 
-    task['targets'].each do |target|
-      template "#{backup_task_dir}/#{task['name']}-#{target}" do
+    task[:targets].each do |target|
+      template "#{backup_task_dir}/#{task[:name]}-#{target}" do
         source "/jobs/backup/drive/backup-drive-target.erb"
         owner username
-        group usr['group']
+        group usr[:group]
         mode 00755
         backup false
         variables(
           :backup_task_dir    => backup_task_dir,
-          :backup_task_name   => task['name'],
+          :backup_task_name   => task[:name],
           :backup_task_drive  => task_drive,
           :backup_task_target => target
         )
       end
 
-      template "#{backup_task_dir}/#{task['name']}-#{target}-excl_patterns.lst" do
+      template "#{backup_task_dir}/#{task[:name]}-#{target}-excl_patterns.lst" do
         source "/jobs/backup/drive/backup-drive-target-excl_patterns.lst.erb"
         owner username
-        group usr['group']
+        group usr[:group]
         mode 00644
         backup false
       end
 
-      home_folders = data_bag_item('global', 'home_folders')
-      incl_patterns_list = home_folders[box['lang']].map { |folder| "#{folder}/***\n" }.join
-      incl_patterns_file = "#{backup_task_dir}/#{task['name']}-#{target}-incl_patterns.lst"
+      user_dirs = data_bag_item('global', 'user_dirs')
+      incl_patterns_list = user_dirs[box[:lang]].map { |folder| "#{folder}/***\n" }.join
+      incl_patterns_file = "#{backup_task_dir}/#{task[:name]}-#{target}-incl_patterns.lst"
 
       file incl_patterns_file do
         content incl_patterns_list
         owner username
-        group usr['group']
+        group usr[:group]
         mode 00644
         backup false
       end